requirements/local.txt
-r base.txt

psycopg2-binary==2.9.5

flake8==6.0.0
pyproject-flake8==6.0.0
black==23.1.0
isort==5.12.0
watchfiles==0.18.1

pytest-django==4.5.2
pytest-factoryboy==2.5.1
Faker==18.3.1
pytest-cov==4.0.0
